Refactor DeviceManager
Inject Astarte Device as dependecy to the DeviceManager.
Use Publisher trait on DeviceManager struct is useful both for not depending
directly on the SDK and for testing reasons.
Move Astarte stuffs in data module.
