+++ src/lib.rs
@@ -18,9 +18,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-use astarte_sdk::builder::AstarteOptions;
 use astarte_sdk::types::AstarteType;
-use astarte_sdk::{registration, Aggregation, Clientbound};
+use astarte_sdk::{Aggregation, Clientbound};
 use log::{debug, info, warn};
 use serde::Deserialize;
 use tokio::sync::mpsc::{Receiver, Sender};
@@ -28,14 +27,11 @@ use tokio::sync::mpsc::{Receiver, Sender};
 use device::DeviceProxy;
 use error::DeviceManagerError;
 
-use crate::astarte::Astarte;
-use crate::data::astarte;
+use crate::data::Publisher;
 use crate::ota::ota_handler::OTAHandler;
-use crate::repository::file_state_repository::FileStateRepository;
-use crate::repository::StateRepository;
 
 mod commands;
-mod data;
+pub mod data;
 mod device;
 pub mod error;
 mod ota;
@@ -44,7 +40,7 @@ mod repository;
 mod telemetry;
 pub mod wrapper;
 
-#[derive(Debug, Deserialize)]
+#[derive(Debug, Deserialize, Clone)]
 pub struct DeviceManagerOptions {
     pub realm: String,
     pub device_id: Option<String>,
@@ -57,56 +53,30 @@ pub struct DeviceManagerOptions {
     pub astarte_ignore_ssl: Option<bool>,
 }
 
-pub struct DeviceManager {
-    sdk: Astarte,
+pub struct DeviceManager<T: Publisher + Clone> {
+    publisher: T,
     //we pass all Astarte event through a channel, to avoid blocking the main loop
     ota_event_channel: Sender<Clientbound>,
     data_event_channel: Sender<Clientbound>,
 }
 
-impl DeviceManager {
-    pub async fn new(opts: DeviceManagerOptions) -> Result<DeviceManager, DeviceManagerError> {
-        let device_id: String = get_device_id(opts.device_id.clone()).await?;
-        let credentials_secret: String = get_credentials_secret(
-            &device_id,
-            &opts,
-            FileStateRepository::new(
-                opts.store_directory.clone(),
-                format!("credentials_{}.json", device_id),
-            ),
-        )
-        .await?;
-
-        let mut sdk_options = AstarteOptions::new(
-            &opts.realm,
-            &device_id,
-            &credentials_secret,
-            &opts.pairing_url,
-        );
-
-        if Some(true) == opts.astarte_ignore_ssl {
-            sdk_options.ignore_ssl_errors();
-        }
-
-        let sdk_options = sdk_options
-            .interface_directory(&opts.interfaces_directory)?
-            .build();
-        info!("Starting");
-
+impl<T: Publisher + Clone + 'static> DeviceManager<T> {
+    pub async fn new(
+        opts: DeviceManagerOptions,
+        publisher: T,
+    ) -> Result<DeviceManager<T>, DeviceManagerError> {
         wrapper::systemd::systemd_notify_status("Initializing");
-        let astarte_client = Astarte::new(&sdk_options).await?;
+        info!("Starting");
 
         let ota_handler = OTAHandler::new(&opts).await?;
 
-        ota_handler
-            .ensure_pending_ota_response(&astarte_client)
-            .await?;
+        ota_handler.ensure_pending_ota_response(&publisher).await?;
 
         let (ota_tx, ota_rx) = tokio::sync::mpsc::channel(1);
         let (data_tx, data_rx) = tokio::sync::mpsc::channel(32);
 
         let device_runtime = Self {
-            sdk: astarte_client,
+            publisher,
             ota_event_channel: ota_tx,
             data_event_channel: data_tx,
         };
@@ -121,7 +91,7 @@ impl DeviceManager {
         mut ota_handler: OTAHandler<'static>,
         mut ota_rx: Receiver<Clientbound>,
     ) {
-        let astarte_client_clone = self.sdk.clone();
+        let astarte_client_clone = self.publisher.clone();
         tokio::spawn(async move {
             while let Some(clientbound) = ota_rx.recv().await {
                 match (
@@ -174,7 +144,7 @@ impl DeviceManager {
 
     pub async fn run(&mut self) {
         wrapper::systemd::systemd_notify_status("Running");
-        let w = self.sdk.device_sdk.clone();
+        let w = self.publisher.clone();
         tokio::task::spawn(async move {
             loop {
                 let systatus = telemetry::system_status::get_system_status().unwrap();
@@ -192,7 +162,7 @@ impl DeviceManager {
         });
 
         loop {
-            match self.sdk.device_sdk.poll().await {
+            match self.publisher.on_event().await {
                 Ok(clientbound) => {
                     debug!("incoming: {:?}", clientbound);
 
@@ -218,7 +188,7 @@ impl DeviceManager {
     }
 
     pub async fn send_initial_telemetry(&self) -> Result<(), DeviceManagerError> {
-        let device = &self.sdk.device_sdk;
+        let device = &self.publisher;
 
         let data = [
             (
@@ -250,7 +220,7 @@ impl DeviceManager {
         }
 
         let disks = telemetry::storage_usage::get_storage_usage()?;
-        for (disk_name, storage) in &disks {
+        for (disk_name, storage) in disks {
             device
                 .send_object(
                     "io.edgehog.devicemanager.StorageUsage",
@@ -263,15 +233,7 @@ impl DeviceManager {
     }
 }
 
-async fn get_device_id(opt_device_id: Option<String>) -> Result<String, DeviceManagerError> {
-    if let Some(device_id) = opt_device_id {
-        Ok(device_id)
-    } else {
-        get_hardware_id_from_dbus().await
-    }
-}
-
-async fn get_hardware_id_from_dbus() -> Result<String, DeviceManagerError> {
+pub async fn get_hardware_id_from_dbus() -> Result<String, DeviceManagerError> {
     let connection = zbus::Connection::system().await?;
     let proxy = DeviceProxy::new(&connection).await?;
     let hardware_id: String = proxy.get_hardware_id("").await?;
@@ -283,154 +245,17 @@ async fn get_hardware_id_from_dbus() -> Result<String, DeviceManagerError> {
     Ok(hardware_id)
 }
 
-async fn get_credentials_secret(
-    device_id: &str,
-    opts: &DeviceManagerOptions,
-    cred_state_repo: impl StateRepository<String>,
-) -> Result<String, DeviceManagerError> {
-    if let Some(secret) = opts.credentials_secret.clone() {
-        Ok(secret)
-    } else if cred_state_repo.exists() {
-        get_credentials_secret_from_persistence(cred_state_repo)
-    } else if let Some(token) = opts.pairing_token.clone() {
-        get_credentials_secret_from_registration(device_id, &token, opts, cred_state_repo).await
-    } else {
-        Err(DeviceManagerError::FatalError(
-            "Missing arguments".to_string(),
-        ))
-    }
-}
-
-fn get_credentials_secret_from_persistence(
-    cred_state_repo: impl StateRepository<String>,
-) -> Result<String, DeviceManagerError> {
-    Ok(cred_state_repo.read().expect("Unable to read secret"))
-}
-
-async fn get_credentials_secret_from_registration(
-    device_id: &str,
-    token: &str,
-    opts: &DeviceManagerOptions,
-    cred_state_repo: impl StateRepository<String>,
-) -> Result<String, DeviceManagerError> {
-    let registration =
-        registration::register_device(token, &opts.pairing_url, &opts.realm, &device_id).await;
-    if let Ok(credentials_secret) = registration {
-        cred_state_repo
-            .write(&credentials_secret)
-            .expect("Unable to write secret");
-        Ok(credentials_secret)
-    } else {
-        Err(DeviceManagerError::FatalError("Pairing error".to_string()))
-    }
-}
-
 #[cfg(test)]
 mod tests {
-    use crate::repository::MockStateRepository;
-    use crate::{
-        get_credentials_secret, get_credentials_secret_from_registration, get_device_id,
-        DeviceManager, DeviceManagerError, DeviceManagerOptions,
-    };
+    use crate::data::MockPublisher;
+    use crate::{DeviceManager, DeviceManagerOptions};
 
-    #[tokio::test]
-    async fn device_id_test() {
-        assert_eq!(
-            get_device_id(Some("target".to_string())).await.unwrap(),
-            "target".to_string()
-        );
-    }
+    impl Clone for MockPublisher {
+        fn clone(&self) -> Self {
+            MockPublisher::new()
+        }
 
-    #[tokio::test]
-    async fn credentials_secret_test() {
-        let state_mock = MockStateRepository::<String>::new();
-        let options = DeviceManagerOptions {
-            realm: "".to_string(),
-            device_id: None,
-            credentials_secret: Some("credentials_secret".to_string()),
-            pairing_url: "".to_string(),
-            pairing_token: None,
-            interfaces_directory: "".to_string(),
-            store_directory: "".to_string(),
-            download_directory: "".to_string(),
-            astarte_ignore_ssl: Some(false),
-        };
-        assert_eq!(
-            get_credentials_secret("device_id", &options, state_mock)
-                .await
-                .unwrap(),
-            "credentials_secret".to_string()
-        );
-    }
-
-    #[tokio::test]
-    async fn not_enough_arguments_credentials_secret_test() {
-        let mut state_mock = MockStateRepository::<String>::new();
-        state_mock.expect_exists().returning(|| false);
-        let options = DeviceManagerOptions {
-            realm: "".to_string(),
-            device_id: None,
-            credentials_secret: None,
-            pairing_url: "".to_string(),
-            pairing_token: None,
-            interfaces_directory: "".to_string(),
-            store_directory: "".to_string(),
-            download_directory: "".to_string(),
-            astarte_ignore_ssl: Some(false),
-        };
-        assert!(get_credentials_secret("device_id", &options, state_mock)
-            .await
-            .is_err());
-    }
-
-    #[tokio::test]
-    #[should_panic(expected = "Unable to read secret: FatalError(\"\")")]
-    async fn get_credentials_secret_persistence_fail() {
-        let mut state_mock = MockStateRepository::<String>::new();
-        state_mock.expect_exists().returning(|| true);
-        state_mock
-            .expect_read()
-            .returning(move || Err(DeviceManagerError::FatalError("".to_owned())));
-
-        let options = DeviceManagerOptions {
-            realm: "".to_string(),
-            device_id: Some("device_id".to_owned()),
-            credentials_secret: None,
-            pairing_url: "".to_string(),
-            pairing_token: None,
-            interfaces_directory: "".to_string(),
-            store_directory: "".to_string(),
-            download_directory: "".to_string(),
-            astarte_ignore_ssl: Some(false),
-        };
-
-        assert!(get_credentials_secret("device_id", &options, state_mock)
-            .await
-            .is_err());
-    }
-
-    #[tokio::test]
-    async fn get_credentials_secret_persistence_success() {
-        let mut state_mock = MockStateRepository::<String>::new();
-        state_mock.expect_exists().returning(|| true);
-        state_mock
-            .expect_read()
-            .returning(move || Ok("cred_secret".to_owned()));
-
-        let options = DeviceManagerOptions {
-            realm: "".to_string(),
-            device_id: Some("device_id".to_owned()),
-            credentials_secret: None,
-            pairing_url: "".to_string(),
-            pairing_token: None,
-            interfaces_directory: "".to_string(),
-            store_directory: "".to_string(),
-            download_directory: "".to_string(),
-            astarte_ignore_ssl: Some(false),
-        };
-        assert!(get_credentials_secret("device_id", &options, state_mock)
-            .await
-            .is_ok());
+        fn clone_from(&mut self, _: &Self) {}
     }
 
     #[tokio::test]
@@ -446,7 +271,7 @@ mod tests {
             download_directory: "".to_string(),
             astarte_ignore_ssl: Some(false),
         };
-        let mut dm = DeviceManager::new(options).await;
+        let dm = DeviceManager::new(options, MockPublisher::new()).await;
 
         assert!(dm.is_err());
     }
@@ -465,36 +290,9 @@ mod tests {
             download_directory: "".to_string(),
             astarte_ignore_ssl: Some(false),
         };
-        let mut dm = DeviceManager::new(options).await;
+        let dm = DeviceManager::new(options, MockPublisher::new()).await;
 
         assert!(dm.is_ok());
-    }
-
-    #[tokio::test]
-    async fn get_credentials_secret_from_registration_fail() {
-        let options = DeviceManagerOptions {
-            realm: "".to_string(),
-            device_id: Some("device_id".to_string()),
-            credentials_secret: Some("credentials_secret".to_string()),
-            pairing_url: "".to_string(),
-            pairing_token: None,
-            interfaces_directory: "./".to_string(),
-            store_directory: "".to_string(),
-            download_directory: "".to_string(),
-            astarte_ignore_ssl: Some(false),
-        };
-        let mut state_mock = MockStateRepository::<String>::new();
-        let cred_result =
-            get_credentials_secret_from_registration("", "", &options, state_mock).await;
-        assert!(cred_result.is_err());
-        match cred_result.err().unwrap() {
-            DeviceManagerError::FatalError(val) => {
-                assert_eq!(val, "Pairing error".to_owned())
-            }
-            _ => {
-                panic!("Wrong DeviceManagerError type");
-            }
-        };
     }
 }
