+++ src/main.rs
@@ -25,6 +25,7 @@ use std::panic::{self, PanicInfo};
 use std::path::Path;
 
 use config::read_options;
+use edgehog_device_runtime::data::astarte::{astarte_map_options, Astarte};
 use edgehog_device_runtime::error::DeviceManagerError;
 
 mod config;
@@ -73,7 +74,9 @@ async fn main() -> Result<(), DeviceManagerError> {
         })?;
     }
 
-    let mut dm = edgehog_device_runtime::DeviceManager::new(options).await?;
+    let astarte_options = astarte_map_options(&options).await?;
+    let astarte = Astarte::new(astarte_options).await?;
+    let mut dm = edgehog_device_runtime::DeviceManager::new(options, astarte).await?;
 
     dm.init().await?;
